+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 **Improvements and fixes**
 - Added source for each alert for compatible sources
 - [Widgets] “Material You - Current” Tap to open app now works again
+- [Widgets] “Material You - Current” automatically adapt size to the frame on Android >= 12
 - [Widgets] “Material You - Current” add preview
 - Fix current location keeping old timezone on some sources, when travelling to a different timezone
 - Fix “Stay informed” card stays visible after allowing notifications via settings / app-info